Adding option to enable pgbouncer

+++ internal/flypg/readonly.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/fly-apps/postgres-flex/internal/flypg/admin"
 	"github.com/fly-apps/postgres-flex/internal/utils"
+	"github.com/jackc/pgx/v5"
 )
 
 const (
@@ -127,9 +128,23 @@ func removeReadOnlyLock() error {
 }
 
 func changeReadOnlyState(ctx context.Context, n *Node, enable bool) error {
-	conn, err := n.NewPrimaryConnection(ctx, "postgres")
-	if err != nil {
-		return fmt.Errorf("failed to establish connection: %s", err)
+	var (
+		err  error
+		conn *pgx.Conn
+	)
+
+	if n.PGBouncerEnabled {
+		conn, err = n.NewPrimaryConnection(ctx, "postgres")
+		if err != nil {
+			return fmt.Errorf("failed to establish connection: %s", err)
+		}
+		defer conn.Close(ctx)
+	} else {
+		conn, err = n.NewLocalConnection(ctx, "postgres")
+		if err != nil {
+			return fmt.Errorf("failed to establish connection: %s", err)
+		}
+		defer conn.Close(ctx)
 	}
 
 	databases, err := admin.ListDatabases(ctx, conn)
@@ -152,32 +167,38 @@ func changeReadOnlyState(ctx context.Context, n *Node, enable bool) error {
 		dbNames = append(dbNames, db.Name)
 	}
 
-	bConn, err := n.PGBouncer.NewConnection(ctx)
-	if err != err {
-		return fmt.Errorf("failed to establish connection to pgbouncer: %s", err)
-	}
-	defer bConn.Close(ctx)
-
-	poolMode, err := n.PGBouncer.poolMode()
-	if err != nil {
-		return fmt.Errorf("failed to resolve active pool mode: %s", err)
-	}
-
-	switch poolMode {
-	case transactionPooler, statementPooler:
-		if err := n.PGBouncer.forceReconnect(ctx, dbNames); err != nil {
-			return fmt.Errorf("failed to force connection reset: %s", err)
+	if n.PGBouncerEnabled {
+		bConn, err := n.PGBouncer.NewConnection(ctx)
+		if err != err {
+			return fmt.Errorf("failed to establish connection to pgbouncer: %s", err)
 		}
-	case sessionPooler:
-		if err := n.PGBouncer.killConnections(ctx, dbNames); err != nil {
-			return fmt.Errorf("failed to kill connections: %s", err)
+		defer bConn.Close(ctx)
+
+		poolMode, err := n.PGBouncer.poolMode()
+		if err != nil {
+			return fmt.Errorf("failed to resolve active pool mode: %s", err)
 		}
 
-		if err := n.PGBouncer.resumeConnections(ctx, dbNames); err != nil {
-			return fmt.Errorf("failed to resume connections: %s", err)
+		switch poolMode {
+		case transactionPooler, statementPooler:
+			if err := n.PGBouncer.forceReconnect(ctx, dbNames); err != nil {
+				return fmt.Errorf("failed to force connection reset: %s", err)
+			}
+		case sessionPooler:
+			if err := n.PGBouncer.killConnections(ctx, dbNames); err != nil {
+				return fmt.Errorf("failed to kill connections: %s", err)
+			}
+
+			if err := n.PGBouncer.resumeConnections(ctx, dbNames); err != nil {
+				return fmt.Errorf("failed to resume connections: %s", err)
+			}
+		default:
+			return fmt.Errorf("failed to resolve valid pooler. found: %s", poolMode)
 		}
-	default:
-		return fmt.Errorf("failed to resolve valid pooler. found: %s", poolMode)
+	} else {
+		if err := utils.RunCommand("restart-haproxy", "root"); err != nil {
+			return fmt.Errorf("failed to restart haproxy: %s", err)
+		}
 	}
 
 	return nil
